README.md
This repository contains a bachelor’s dissertation titled "The Dark Side of Capitalism: The Slave Trade in Britain from the 17th to the 19th Century." The dissertation explores the intricate relationship between capitalism and the slave trade in Britain, drawing heavily on Eric Williams' influential work Capitalism and Slavery.

Written in LaTeX using Visual Studio Code, this document examines how capitalist economic practices were intertwined with and bolstered by the slave trade over three centuries. It presents a detailed historical analysis supported by extensive research and primary sources.

The project is structured to facilitate easy navigation and compilation, with separate directories for chapters, sections, and supplementary materials. The LaTeX template used for this dissertation is adapted from Luca Modica’s Bachelor Thesis Template on Overleaf.

This repository includes all necessary files to compile the dissertation into a final PDF document and is intended for academic use and further research on the subject.